requirements.txt
Flask
Flask_Login
Flask-SQLAlchemy
Flask-WTF
python-dotenv
psycopg2-binary
gunicorn
